improvement(deps): remove dependency on fastify and @types/express
By removing the hard dependnecy on the types FastifyRequest, Request, FastifyReply<ServerResponse>,
and Response, the fastify and @types/express packages are no longer hard dependencies which will cut
down on the package size. To get around this, minimum types were created (or used) to make sure that
there could still be a distinction between Fastify and Express requests.

fix #4