feat(module): implements plugin system for interceptor context parser
Oh where to begin? This rewrite allows for the OgmaModule to manage so much more than just HTTP, and
in such a way that it's almost hard to describe. With the plugin system created, so long as a class
extends the `AbstractInterceptorService` **and** the class does not have any extra injections
needed, any class can be provided to the `interceptor` property of the `OgmaModuleOptions` to allow
for interceptor request parsing. By default, the class that is provided is a
`NoopInterceptorService` that really should only be there so an error gets thrown, but something had
to be there for the Nest DI ssytem to not complain. Tests are still to come to help clean out dead
code, but overall, the proof of concept is there and it works!

re #7 #8 #9 #10 #11