feat(socket.io): implements parser for socket.io
The SocketioInterceptorParser has two hardcoded values, one for method and one for protocol. It also
only has the concept of pass (200) and fail (500), though there could be a case for checking if the
error is an `HttpException` and if so getting the status from there like is done in the HttpParsers.

fix #22