feat(tcp): implements first draft of tcp parser
The TCP parser pulls information from the socket connection and has a hard codeded method of TCP,
though this may change to protocol instead of method. Protocol currently takes from the remote
connections family (an IP version). The call point must be stringified in the case of using a json
as the message pattern (as the NestJS docs show). Integration has also been built for this, using
concurrently to run the client and server servers.

fix #23