style(all): moves tests to separate folder in each package
There is now a pre and post build script that clears the `lib` directory and then moves everything
from `lib/src` to `lib/` and removes the `lib/src` directory to ensure that when publishing
`lib/index.js` is still populated properly. All tests now live in the `test` directory and source
code in the `src` dir. This will help keep a cleaner folder strucutre and not need to look through
source and spec files when working on new developments.