feat(gql-fastify): implements parser for GraphQL-Fastify
Due to the nature of how apollo-server-fastify works, the `getStatus` method is unusable from any of
the previous package implementations. Instead of extending and making this package require
`platform-express` or `platform-fastify` due to how the `req` object is added, all the methods are
re-implemented for ease. Small code duplication to keep the package size low.

fix #15