perf(logger): reimplements string log method to use template over array
Using template literals brings a small increase in speed to the logger over using `array.join()`.
Also adds more stability and predictability when it comes to use (in my opinion). This also adds the
`@ogma/logger` to the CI test runs, as it was missing a `test:cov` script before now.