fix(logger): error objects are now printed instead {}
`Error`s in JavaScript are Objects with non-enumerable properties
meaning that `JSON.stringify()` returns `{}`. With the new fix
errors no longer are stringified and printed how they are by default
printed when using `process.stdout.write()`.

```ts
process.stdout.write(new Error('This is an error');
// Error: This is an error
```

This should make for cleaner logic and better integration with the
NestJS framework.

Also, the assignment of `this.pid` and `this.hostname` has been
moved up in the constructor to accomodate bad log levels. PID is now
also wrapped in `[]` instead of printed directly, to be more
consistent with the stream logger.

fix #526