feat(styler): creates styler package
The styler package is very similar to chalk in that it adds
a bunch of styling methods to an easy to use API, things like
`bold` and `underline` along with the seven base colors
`black`, `red`, `green`, `yellow`, `blue`,
`magenta`, `cyan` and `white` along with their bright
variants for both foreground (text) and background codes.

Some tests still need to be added, but generally this will
allow for all sorts of colorizing  to happen and to allow
for very pretty logs and terminal messages. There needs to
be a little bit more logic in terms of allowing for colors
that are outside the current streams color depth, and
determinging what needs to happen in those cases,
but generally the code works as intended.