chore: remove workspaces from package.json
If there is a `workspaces` property in the `package.json`
then `@manypkg/get-packages` considers `yarn` to be the
primary package manager, meaning that the the `pnpm-workspace`
file is not checked for where to find the packages, and the pre-
publish step in my `release.yml` action was not being taken into
account. I love that open source allows me to go find that this is
the case, but man was this interesting to find.

[related code](https://github.com/Thinkmill/manypkg/blob/master/packages/get-packages/src/index.ts\#L45)