feat(module): create a decorator for auto logging
There is a new `@Log()` decorator that can now be used to auto
log the start and stop of a function. Currently it is set to use

`log` but I'll make it configurable to use `debug` or something
else, possibly calling an `OgmaService` method to make it
configurable via the `OgmaModule.forRoot()` which would be really
powerful. I also want to add a `@LogAll()` that applies the
`@Log()` decorator to all of the methods of the class. These
are both additions to copme after finishing the `uvu` migration.