chore: setup new coverage collection system
We'll now run all tests twice, which is a little inefficient
but it allows us to see where singular failures are and if
there are none, then collect coverage. As it only takes a literal
second to run all of the tests again, I'm nt worries about
CI performance here.