chore: update package.json of several packages
With Nx 15, the dependencies of each package needs to be  set in the
package.json so that Nx can properly know what is a library to pull from
the build cache and what should be bundled in with the currently being
built library. Once the package.json's were updated, all the builds
succeeded as expected.