docs: convert docs to use  astro
The big reason for using astro over docusuarus is because docu requires
the use of  wait-on which requires `axios` still back on 0.26.x, which is
incompatible with the newest version of Nx so the docs cannot be built or
deployed. This is a **major** problem when it comes to actually developing
on the project because new documentation cannot be published.

I've done what I can at the moment to make the docs feel as much like they
used to whle still using the newer framework.