feat: rewrite the parser gathering logic
Instead of forcing the user to register one of the four paresers and being
stuck with only those contexts, users can now add parers to a `providers`
array and let the `DiscoveryService` find the parsers and then use the
correct one based on the context type. This allows for custom contexts
like necord's `discord` context, to have parsers created and used for it
without having to change the context or assume a certain type for it.

Docs, changesets, and extra tests to come soon

related: #1451