fix: allow for the value of a masked property to be undefined or null
Now when using the `masks` option, if an object comes in with an `undefined`
or `null` value, instead of throwing an error or a message about being too
complex to serialize, Ogma will properly handle the value and replace it with
nine (9) asterisks as a default. This will probably eventually be made
configurable if the desire is there. Alternatively, we may end up just
skipping the masking instead. Will wait for feedback

ref: #1678